Seamless-Updates: renaming latest-mac.yml file in artifactBuildCompleted phase
