+++ packages/app-desktop/afterAllArtifactBuild.js
@@ -5,21 +5,6 @@ const sha512 = require('js-sha512');
 const distDirName = 'dist';
 const distPath = path.join(__dirname, distDirName);
 
-const renameLatestYmlFile = () => {
-	if (os.platform() === 'darwin' && process.arch === 'arm64') {
-		const latestMacFilePath = path.join(distPath, 'latest-mac.yml');
-		const renamedMacFilePath = path.join(distPath, 'latest-mac-arm64.yml');
-		const filesInDist = fs.readdirSync(distPath);
-
-		if (fs.existsSync(latestMacFilePath)) {
-			fs.renameSync(latestMacFilePath, renamedMacFilePath);
-			return [renamedMacFilePath];
-		} else {
-			throw new Error('latest-mac.yml not found! Debugging: Files in dist: ', filesInDist);
-		}
-	}
-};
-
 const generateChecksumFile = () => {
 	if (os.platform() !== 'linux') {
 		return []; // SHA-512 is only for AppImage
@@ -47,7 +32,6 @@ const generateChecksumFile = () => {
 };
 
 const mainHook = () => {
-	renameLatestYmlFile();
 	generateChecksumFile();
 };
